Improve timeout middleware tests, and move them to a separate file
The timeout middleware tests didn't actually test the middleware, just the
address API, which isn't enough. These tests tests the entire chain, so we
can tell what happens when an invalid value is passed.

The tests were also moved into their own file.
